Push to 5.0:
a) During CSI reassignment from active to standby, check if the SI to which the CSI belongs is within the configured limits of SG's activeSUsPerSI. This check should also indirectly fix random failures of amfOnePlusTwo test case during an autoadjust on a 1 active + 2 standby SG with spare.

b) Isolate the m+n specific stuffs to a separate file to have per redundancy implementation specific files.



git-svn-id: http://clovisforge.openclovis.org:8888/svn/clasp/branches/5.0@5493 8a546835-ab24-0410-9070-dd6b907a2c58