Cleanups to xport/multi-xport layer now using nodenames to talk to destinations. Cache the local nodename for faster lookups, no need to invoke map find and map update since map update has a find. So realign the map update and the caller implementation accordingly. Search the nodemap entry based on complete nodenames so we don't have a match between WorkerI11 and WorkerI1.
Also the new nodename xport map has a major flaw in the sense that node names/versions and capability fetched through nodecache now are exchanged as a notification packet in response to receiving a node arrival packet. So the initial notification/discovery itself would fail since node 1 on receiving arrival of node 2(mcast discovery or automated tipc topology) would try to unicast the response back to the new node and that would fail as slot 2 nodename map wouldn't exist in the first place as its received as a response to the notification map of {version,name,capabilities} from slot 1 and hence both would fail trying to unicast the discovery or capability/nodename/version to each other thereby failing the multi-xport protocol. Since this involves an initial discovery and exchange, its advisable to unicast the response to the node using the transport in which the packet was received in the first place (thats why my clIocSendWithXport changes were done and used by notification). So a failure of clFindTransport to look up the map can fall back to the preferredXport if passed so the nodes can initially exchange their capabilities and maps as otherwise the discovery protocol itself would be broken as it talks over slot to exchange nodename/capabilities/versions and that wouldn't work with the new realignment.