Using mgtGet("/module") does not work because the module was not "bound" (registered -- the contained objects were each individually.  This is a problem because the checkpoint resolves them directly so there is no "top level" /module representation, except that which can be inferred from the contents of the checkpoint.
Its also a more intuitive and simpler API to just register the module rather then every contained object.  Note though that you CAN still register individual objects if you need different implementers handling a single module -- but we are going to have to enhance the client side search through the checkpoint to handle the wildcard case given individual registrations.

The contained objects needed to be registered individually because of a object hierarchy that did not follow that defined by the YANG file.  The Module object as created by hand, and the generated code created a separate Container object that contained the top level objects.  So these were siblings when they should have been a single unified object.

To fix this, the Module object now derives from Container so a single object is all that is needed.  This mirrors the YANG definition file so is more intuitive.

Management xpaths can be preceded with enhanced functionality in braces eg: {}.  The first enhanced functionality was added which is to basically "step" into a mgt call by allow a breakpoint {b} or thread pause {p} to happen on the server side.  In this case the client will also never retry the send.

Changes were also made to resolvePath to handle /module and to handle .../list/list[key="value"] xpath requests.

NOTE: The code generator needs to be changed to derive the generated top level class from Module.  This change has been made by hand in SAFplusAMF